+++ tests/sortviz/test_utils.py
@@ -1,7 +1,9 @@
 import json
 import os
 import pathlib
+from typing import List
 
+import matplotlib
 import matplotlib.pyplot as plt
 import numpy as np
 import pytest
@@ -43,6 +45,7 @@ def test_generate_data(num_rows: int, num_cols: int):
                         m_utils.PassDataModel(num_pass=0, row=0, col=1, value=2),
                     ]
                 ],
+                comparisons=m_utils.ComparisonModel(comparisons={0: 10, 1: 20}),
             ),
             "fun_file.txt",
         )
@@ -54,6 +57,7 @@ def test_plot_sort(
     data: m_utils.SortDataModel,
     filename: str,
 ):
+    matplotlib.use("Agg")
     sort_filename = tmp_path / filename
     parse_spy = mocker.spy(m_utils.SortDataModel, "parse_file")
     plot_mock = mocker.patch.object(plt.Axes, "imshow")
@@ -73,6 +77,7 @@ def test_plot_sort(
 
 
 def test_plot_sort_invalid_filename(mocker: pytest_mock.MockerFixture):
+    matplotlib.use("Agg")
     parse_spy = mocker.spy(m_utils.SortDataModel, "parse_file")
     plot_mock = mocker.patch.object(plt.Axes, "imshow")
 
@@ -84,3 +89,59 @@ def test_plot_sort_invalid_filename(mocker: pytest_mock.MockerFixture):
 
     parse_spy.assert_not_called()
     plot_mock.assert_not_called()
+
+
+@pytest.mark.parametrize(
+    "data, filenames",
+    [
+        (
+            m_utils.SortDataModel(
+                method="test_method",
+                rows=2,
+                cols=2,
+                passes=[
+                    [
+                        m_utils.PassDataModel(num_pass=0, row=0, col=0, value=1),
+                        m_utils.PassDataModel(num_pass=0, row=0, col=1, value=2),
+                    ]
+                ],
+                comparisons=m_utils.ComparisonModel(comparisons={0: 10, 1: 20}),
+            ),
+            ["fun_file2.txt"],
+        )
+    ],
+)
+def test_plot_sort_comparison(
+    mocker: pytest_mock.MockerFixture,
+    tmp_path: pathlib.Path,
+    data: m_utils.SortDataModel,
+    filenames: List[str],
+) -> None:
+    """Tests that the plot_sort_comparison will create a bar plot with the provided averages.
+
+    :param mocker: a mocker test fixture.
+    :type mocker: pytest_mock.MockerFixture
+    :param tmp_path: a temp path to save data to.
+    :type tmp_path: pathlib.Path
+    :param data: the data to plot
+    :type data: m_utils.SortDataModel
+    :param filenames: the filenames containing data to compare.
+    :type filenames: List[str]
+    """
+    matplotlib.use("Agg")
+    sort_filenames = [str(tmp_path / _file) for _file in filenames]
+    parse_spy = mocker.spy(m_utils.SortDataModel, "parse_file")
+    plot_mock = mocker.patch.object(plt.Axes, "bar")
+    num_sorters = len(filenames)
+    # create temp data
+    for sort_filename in sort_filenames:
+        with open(sort_filename, "w") as f:
+            json.dump(data.dict(), f, ensure_ascii=True)
+
+    m_utils.plot_sort_comparison(sort_filenames)
+
+    parse_spy.assert_called_once_with(sort_filenames[0])
+    plot_mock.assert_called_once()
+
+    assert np.array_equal(plot_mock.call_args.args[0], np.arange(num_sorters))
+    assert np.array_equal(plot_mock.call_args.args[1], [np.mean([10, 20])])  # averages